requirements.txt

discord.py==1.6.0
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
psycopg2==2.8.6
